Refactor: 이미지 관련 로직 PR 반영
