+++ src/main/java/com/studypals/domain/common/fileManage/FilePath.java
@@ -1,4 +1,4 @@
-package com.studypals.domain.imageManage.dto;
+package com.studypals.domain.common.fileManage;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 /**
- * 도메인 별 이미지 저장 디렉토리를 정의합니다.
+ * 도메인 별 파일 저장 디렉토리를 정의합니다.
  *
  * <p>enum으로 직접적인 경로 디렉토리를 관리합니다.
  *
@@ -15,8 +15,8 @@ import lombok.RequiredArgsConstructor;
  */
 @Getter
 @RequiredArgsConstructor
-public enum ImagePath {
-    USER("user", List.of("jpg", "jpeg", "png", "bmp", "webp"));
+public enum FilePath {
+    USER_PROFILE_IMAGE("user-profile", List.of("jpg", "jpeg", "png", "bmp", "webp"));
 
     private final String path;
     private final List<String> acceptableExtensions;
